+++ src/types/resolvers.ts
@@ -2,7 +2,7 @@ import { ParamDefinition } from "../metadata/definition-interfaces";
 
 export interface BaseResolverDefinitions {
   methodName: string;
-  handler: Function;
   target: Function;
+  handler?: Function;
   params?: ParamDefinition[];
 }