+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - add support for ommiting return type when use type options, in selected decorators (`@Field`, `@Arg`)
 
+#### Fixes
+
+- fix class getter resolvers bug - missing fields from prototype (`plainToClass` bug)
+
 ### 0.1.0
 
 #### Initial release